requirements.txt
asgiref==3.8.1
Django==4.2.23
django-ckeditor==6.7.3
django-js-asset==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
